refactor: ApprovalChainBuilder を domain/services/approval/ へ区画化 (#421)
承認連鎖ビルダーと隣接テストを domain/services/approval/ へ集約。
本体・テストの values 参照を +1 階層調整、自テスト ../ApprovalChainBuilder は
共移動のため不変。参照元は自テストのみ。build/eslint(estimate)/test green。
